review: cleanup ulong test types and receipt-index check
JsonRpcServiceTests: align Number/Size TestCase to ulong, drop (long)
casts. TraceRpcModuleTests: ModExpGasUsed returns ulong. FeeHistoryOracleTests:
maxDistFromHead ulong?, drop redundant cast. Eth70ProtocolHandler: replace
opaque (ulong)/(uint) coercion with explicit negative + range check.

+++ src/Nethermind/Nethermind.JsonRpc.Test/JsonRpcServiceTests.cs
@@ -206,15 +206,15 @@ public class JsonRpcServiceTests
         return response;
     }
 
-    [TestCase(false, 2L, TestName = "Number")]
-    [TestCase(true, 513L, TestName = "Size")]
-    public void Eth_module_populates_block_data(bool assertSize, long expected)
+    [TestCase(false, 2UL, TestName = "Number")]
+    [TestCase(true, 513UL, TestName = "Size")]
+    public void Eth_module_populates_block_data(bool assertSize, ulong expected)
     {
         IEthRpcModule ethRpcModule = Substitute.For<IEthRpcModule>();
         ISpecProvider specProvider = Substitute.For<ISpecProvider>();
         ethRpcModule.eth_getBlockByNumber(Arg.Any<BlockParameter>(), true).ReturnsForAnyArgs(x => ResultWrapper<BlockForRpc>.Success(new BlockForRpc(Build.A.Block.WithNumber(2).TestObject, true, specProvider)));
         BlockForRpc result = RpcTest.AssertSuccess<BlockForRpc>(TestRequest(ethRpcModule, "eth_getBlockByNumber", "0x1b4", "true"));
-        Assert.That(assertSize ? (long)result.Size : (long)result.Number!.Value, Is.EqualTo(expected));
+        Assert.That(assertSize ? (ulong)result.Size : result.Number!.Value, Is.EqualTo(expected));
     }
 
     [Test]
